Move all Analytics Hub classes to specific package
