+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/hub/AnalyticsHubViewState.kt
@@ -1,8 +1,8 @@
-package com.woocommerce.android.ui.analytics
+package com.woocommerce.android.ui.analytics.hub
 
-import com.woocommerce.android.ui.analytics.daterangeselector.AnalyticsHubDateRangeSelectorViewState
-import com.woocommerce.android.ui.analytics.informationcard.AnalyticsHubInformationViewState
-import com.woocommerce.android.ui.analytics.listcard.AnalyticsHubListViewState
+import com.woocommerce.android.ui.analytics.hub.daterangeselector.AnalyticsHubDateRangeSelectorViewState
+import com.woocommerce.android.ui.analytics.hub.informationcard.AnalyticsHubInformationViewState
+import com.woocommerce.android.ui.analytics.hub.listcard.AnalyticsHubListViewState
 import com.woocommerce.android.viewmodel.MultiLiveEvent
 
 data class AnalyticsViewState(
