+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/hub/listcard/AnalyticsHubListCardView.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.analytics.listcard
+package com.woocommerce.android.ui.analytics.hub.listcard
 
 import android.content.Context
 import android.util.AttributeSet
@@ -11,9 +11,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.google.android.material.card.MaterialCardView
 import com.woocommerce.android.R
 import com.woocommerce.android.databinding.AnalyticsListCardViewBinding
-import com.woocommerce.android.ui.analytics.listcard.AnalyticsHubListViewState.DataViewState
-import com.woocommerce.android.ui.analytics.listcard.AnalyticsHubListViewState.LoadingViewState
-import com.woocommerce.android.ui.analytics.listcard.AnalyticsHubListViewState.NoDataState
+import com.woocommerce.android.ui.analytics.hub.listcard.AnalyticsHubListViewState.DataViewState
+import com.woocommerce.android.ui.analytics.hub.listcard.AnalyticsHubListViewState.LoadingViewState
+import com.woocommerce.android.ui.analytics.hub.listcard.AnalyticsHubListViewState.NoDataState
 import com.woocommerce.android.widgets.SkeletonView
 import com.woocommerce.android.widgets.tags.ITag
 import com.woocommerce.android.widgets.tags.TagConfig
