+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/hub/sync/UpdateAnalyticsHubStats.kt
@@ -1,13 +1,13 @@
-package com.woocommerce.android.ui.analytics.sync
+package com.woocommerce.android.ui.analytics.hub.sync
 
 import com.woocommerce.android.model.OrdersStat
 import com.woocommerce.android.model.ProductsStat
 import com.woocommerce.android.model.RevenueStat
 import com.woocommerce.android.model.SessionStat
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubDateRangeSelection
-import com.woocommerce.android.ui.analytics.sync.AnalyticsHubUpdateState.Finished
-import com.woocommerce.android.ui.analytics.sync.AnalyticsHubUpdateState.Loading
-import com.woocommerce.android.ui.analytics.sync.AnalyticsRepository.FetchStrategy
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsHubUpdateState.Finished
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsHubUpdateState.Loading
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsRepository.FetchStrategy
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.combine
