+++ WooCommerce/src/main/res/layout/analytics_information_card_view.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".ui.analytics.informationcard.AnalyticsHubInformationCardView">
+    tools:context=".ui.analytics.hub.informationcard.AnalyticsHubInformationCardView">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/analyticsCardDataContainer"
@@ -38,7 +38,7 @@
             tools:text="Sales"
             tools:visibility="visible" />
 
-        <com.woocommerce.android.ui.analytics.informationcard.AnalyticsHubInformationSectionView
+        <com.woocommerce.android.ui.analytics.hub.informationcard.AnalyticsHubInformationSectionView
             android:id="@+id/leftAnalyticsSection"
             style="@style/Woo.Card.WithoutPadding"
             android:layout_width="0dp"
@@ -51,7 +51,7 @@
             app:layout_constraintWidth_percent="0.5"
             tools:visibility="visible" />
 
-        <com.woocommerce.android.ui.analytics.informationcard.AnalyticsHubInformationSectionView
+        <com.woocommerce.android.ui.analytics.hub.informationcard.AnalyticsHubInformationSectionView
             android:id="@+id/rightAnalyticsSection"
             style="@style/Woo.Card.WithoutPadding"
             android:layout_width="0dp"
