+++ WooCommerce/src/main/res/layout/analytics_list_card_item.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.woocommerce.android.ui.analytics.listcard.AnalyticsHubListCardItemView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.woocommerce.android.ui.analytics.hub.listcard.AnalyticsHubListCardItemView xmlns:android="http://schemas.android.com/apk/res/android"
     style="@style/Woo.Card.WithoutPadding"
     android:layout_width="match_parent"
     android:layout_height="wrap_content" />
