+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/analytics/AnalyticsHubViewModelTest.kt
@@ -9,20 +9,25 @@ import com.woocommerce.android.model.ProductItem
 import com.woocommerce.android.model.ProductsStat
 import com.woocommerce.android.model.RevenueStat
 import com.woocommerce.android.model.SessionStat
-import com.woocommerce.android.ui.analytics.RefreshIndicator.NotShowIndicator
-import com.woocommerce.android.ui.analytics.informationcard.AnalyticsHubInformationSectionViewState
-import com.woocommerce.android.ui.analytics.informationcard.AnalyticsHubInformationViewState
-import com.woocommerce.android.ui.analytics.informationcard.AnalyticsHubInformationViewState.LoadingViewState
-import com.woocommerce.android.ui.analytics.listcard.AnalyticsHubListViewState
+import com.woocommerce.android.ui.analytics.hub.AnalyticsHubTransactionLauncher
+import com.woocommerce.android.ui.analytics.hub.AnalyticsHubViewModel
+import com.woocommerce.android.ui.analytics.hub.AnalyticsViewEvent
+import com.woocommerce.android.ui.analytics.hub.AnalyticsViewState
+import com.woocommerce.android.ui.analytics.hub.RefreshIndicator
+import com.woocommerce.android.ui.analytics.hub.RefreshIndicator.NotShowIndicator
+import com.woocommerce.android.ui.analytics.hub.informationcard.AnalyticsHubInformationSectionViewState
+import com.woocommerce.android.ui.analytics.hub.informationcard.AnalyticsHubInformationViewState
+import com.woocommerce.android.ui.analytics.hub.informationcard.AnalyticsHubInformationViewState.LoadingViewState
+import com.woocommerce.android.ui.analytics.hub.listcard.AnalyticsHubListViewState
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubDateRangeSelection.SelectionType.LAST_YEAR
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubDateRangeSelection.SelectionType.TODAY
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubDateRangeSelection.SelectionType.WEEK_TO_DATE
-import com.woocommerce.android.ui.analytics.sync.AnalyticsHubUpdateState
-import com.woocommerce.android.ui.analytics.sync.OrdersState
-import com.woocommerce.android.ui.analytics.sync.ProductsState
-import com.woocommerce.android.ui.analytics.sync.RevenueState
-import com.woocommerce.android.ui.analytics.sync.SessionState
-import com.woocommerce.android.ui.analytics.sync.UpdateAnalyticsHubStats
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsHubUpdateState
+import com.woocommerce.android.ui.analytics.hub.sync.OrdersState
+import com.woocommerce.android.ui.analytics.hub.sync.ProductsState
+import com.woocommerce.android.ui.analytics.hub.sync.RevenueState
+import com.woocommerce.android.ui.analytics.hub.sync.SessionState
+import com.woocommerce.android.ui.analytics.hub.sync.UpdateAnalyticsHubStats
 import com.woocommerce.android.util.CurrencyFormatter
 import com.woocommerce.android.util.locale.LocaleProvider
 import com.woocommerce.android.viewmodel.BaseUnitTest
