+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/analytics/UpdateAnalyticsHubStatsTest.kt
@@ -4,17 +4,17 @@ import com.woocommerce.android.model.OrdersStat
 import com.woocommerce.android.model.ProductsStat
 import com.woocommerce.android.model.RevenueStat
 import com.woocommerce.android.model.SessionStat
-import com.woocommerce.android.ui.analytics.sync.AnalyticsRepository
-import com.woocommerce.android.ui.analytics.sync.AnalyticsRepository.FetchStrategy.Saved
-import com.woocommerce.android.ui.analytics.sync.AnalyticsRepository.OrdersResult
-import com.woocommerce.android.ui.analytics.sync.AnalyticsRepository.ProductsResult
-import com.woocommerce.android.ui.analytics.sync.AnalyticsRepository.RevenueResult
-import com.woocommerce.android.ui.analytics.sync.AnalyticsRepository.VisitorsResult
-import com.woocommerce.android.ui.analytics.sync.OrdersState
-import com.woocommerce.android.ui.analytics.sync.ProductsState
-import com.woocommerce.android.ui.analytics.sync.RevenueState
-import com.woocommerce.android.ui.analytics.sync.SessionState
-import com.woocommerce.android.ui.analytics.sync.UpdateAnalyticsHubStats
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsRepository
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsRepository.FetchStrategy.Saved
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsRepository.OrdersResult
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsRepository.ProductsResult
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsRepository.RevenueResult
+import com.woocommerce.android.ui.analytics.hub.sync.AnalyticsRepository.VisitorsResult
+import com.woocommerce.android.ui.analytics.hub.sync.OrdersState
+import com.woocommerce.android.ui.analytics.hub.sync.ProductsState
+import com.woocommerce.android.ui.analytics.hub.sync.RevenueState
+import com.woocommerce.android.ui.analytics.hub.sync.SessionState
+import com.woocommerce.android.ui.analytics.hub.sync.UpdateAnalyticsHubStats
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.launchIn
